+++ src/Genius/Resources/ReferentsResource.php
@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 
 namespace Genius\Resources;
 
@@ -6,14 +7,29 @@ namespace Genius\Resources;
  * Class ReferentsResource
  * @package Genius\Resources
  *
- * @see     https://docs.genius.com/#referents-h2
+ * @see https://docs.genius.com/#referents-h2
  */
 class ReferentsResource extends AbstractResource
 {
-    
-    public function get(array $data)
+    public function get(
+        ?int $created_by_id = null,
+        ?int $song_id = null,
+        ?int $web_page_id = null,
+        string $text_format = 'dom',
+        ?int $per_page = null,
+        ?int $page = null
+    ): \stdClass
     {
-        return $this->sendRequest('GET', 'referents/?' . http_build_query($data));
+        return $this->sendRequest(
+            'GET',
+            'referents/?' . http_build_query([
+                'created_by_id' => $created_by_id,
+                'song_id' => $song_id,
+                'web_page_id' => $web_page_id,
+                'text_format' => $text_format,
+                'per_page' => $per_page,
+                'page' => $page,
+            ])
+        );
     }
-    
 }