+++ src/Genius/Resources/ResourceException.php
@@ -1,14 +1,11 @@
 <?php
-/**
- * Created by PhpStorm.
- * User: simivar
- * Date: 2017-06-18
- * Time: 21:28
- */
+declare(strict_types=1);
 
 namespace Genius\Resources;
 
-class ResourceException extends \Exception
+use Genius\GeniusException;
+
+final class ResourceException extends GeniusException
 {
     
 }