+++ src/Genius/Resources/SearchResource.php
@@ -1,4 +1,6 @@
 <?php
+declare(strict_types=1);
+
 namespace Genius\Resources;
 
 /**
@@ -9,10 +11,8 @@ namespace Genius\Resources;
  */
 class SearchResource extends AbstractResource
 {
-    
-    public function get($query)
+    public function get(string $query): \stdClass
     {
         return $this->sendRequest('GET', 'search/?' . http_build_query(['q' => $query]));
     }
-    
 }