+++ src/Genius/Resources/WebPagesResource.php
@@ -1,4 +1,6 @@
 <?php
+declare(strict_types=1);
+
 namespace Genius\Resources;
 
 /**
@@ -9,10 +11,15 @@ namespace Genius\Resources;
  */
 class WebPagesResource extends AbstractResource
 {
-    
-    public function get(array $data)
+    public function get(string $raw_annotatable_url, ?string $canonical_url = null, ?string $og_url = null): \stdClass
     {
-        return $this->sendRequest('GET', 'web_pages/lookup/?' . http_build_query($data));
+        return $this->sendRequest(
+            'GET',
+            'web_pages/lookup/?' . http_build_query([
+                'raw_annotatable_url' => $raw_annotatable_url,
+                'canonical_url' => $canonical_url,
+                'og_url' => $og_url,
+            ])
+        );
     }
-    
 }