+++ CHANGELOG.md
@@ -4,6 +4,39 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
+## [2.0.0] - 2019-12-27
+
+### Added
+
+- Support for HTTPlug 2.0
+- Type-hints everywhere
+- Introduced `Genius\GeniusException`
+
+### Changed
+
+- Signature of `Genius\Resources\ArtistsResource::getSongs` changed from array as second argument
+  to `getSongs(int $id, string $sort = 'title', ?int $per_page = null, ?int $page = null)`
+- Signature of `Genius\Resources\WebPagesResource::get` changed from array as first argument
+  to `get(string $raw_annotatable_url, ?string $canonical_url = null, ?string $og_url = null)`
+- Signature of `Genius\Resources\AnnotationsResource::post` changed from array as first argument
+  to `post(array $annotation, array $referent, array $web_page)`
+- Signature of `Genius\Resources\AnnotationsResource::put` changed from array as first argument
+  to `put(int $id, array $annotation, array $referent, array $web_page)`
+- Signature of `Genius\Resources\ReferentsResource::put` changed from array as first argument
+  to `get(?int $created_by_id = null, ?int $song_id = null, ?int $web_page_id = null, string 
+  $text_format = 'dom', ?int $per_page = null, ?int $page = null)`
+- `Genius\Authentication\OAuth2::getAccessToken` returns `null` instead of `false`
+- `Genius\Authentication\OAuth2::refreshToken` returns `null` instead of `false`
+- All Exceptions thrown by library are children of `Genius\GeniusException`
+- Resources throw `Genius\Resources\ResourceException` when calling a method that requires 
+  `scope` and `Bearer` authentication is used
+- Exception messages in `Genius\Resources namespace` are now more developer-friendly contain 
+  called method name and class
+
+### Removed
+
+- Support for PHP <7.1
+
 ## [1.0.1] - 2017-09-13
 
 ### Added
@@ -23,5 +56,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Initial release
 
+[2.0.0]: https://github.com/simivar/Genius-PHP/compare/1.0.1...2.0.0
 [1.0.1]: https://github.com/simivar/Genius-PHP/compare/1.0.0...1.0.1
 [1.0.0]: https://github.com/simivar/Genius-PHP/compare/0.1.0...1.0.0